+++ packages/pwa-kit-react-sdk/src/ssr/server/express.js
@@ -153,10 +153,11 @@ export const createApp = (options) => {
         // Suppress SSL checks - can be used for local dev server
         // test code. Undocumented at present because there should
         // be no use-case for SDK users to set this.
-        strictSSL: true
+        strictSSL: true,
+
+        enableLegacyRemoteProxying: true
     }
 
-    // Build the options, filling in default values
     options = Object.assign({}, defaults, options)
 
     setQuiet(options.quiet || process.env.SSR_QUIET)
@@ -196,7 +197,6 @@ export const createApp = (options) => {
     // Configure the server with the basic options
     updatePackageMobify(options.mobify)
 
-    // Set up the proxies
     configureProxyConfigs(options.appHostname, options.protocol)
 
     const app = createExpressApp(options)
@@ -248,14 +248,28 @@ export const createApp = (options) => {
         })
     }
 
-    // The /mobify/ping path provides a very fast and lightweight
-    // healthcheck response.
-    app.get('/mobify/ping', (req, res) =>
+    // Healthcheck
+    app.get('/mobify/ping', (_, res) =>
         res
             .set('cache-control', NO_CACHE)
             .sendStatus(200)
             .end()
     )
+
+    // Proxying
+    const shouldProxy = !isRemote() || options.enableLegacyRemoteProxying
+    if (shouldProxy) {
+        proxyConfigs.forEach((config) => {
+            app.use(config.proxyPath, config.proxy)
+            app.use(config.cachingPath, config.cachingProxy)
+        })
+    } else {
+        app.all('/mobify/proxy/*', (_, res) => {
+            return res.status(501).json({
+                error: 'Environment proxies are not set: https://sfdc.co/cc-mrt-proxy-setup'
+            })
+        })
+    }
 
     // Both local and remote modes can perform proxying. A remote
     // server usually only does proxying for development deployments
@@ -391,6 +405,14 @@ const validateConfiguration = (options) => {
                 'in PEM format, whose name ends with ".pem". ' +
                 'See the "cert" and "key" options on ' +
                 'https://nodejs.org/api/tls.html#tls_tls_createsecurecontext_options'
+        )
+    }
+
+    if (options.enableLegacyRemoteProxying) {
+        console.warn(
+            'Legacy proxying behaviour is enabled. ' +
+                'This behaviour is deprecated and will be removed in the future.' +
+                'To disable it, set `options.enableLegacyRemoteProxying: false`.'
         )
     }
 }
@@ -1656,9 +1678,6 @@ const applyPatches = once((options) => {
     // Patch the http.request/get and https.request/get
     // functions to allow us to intercept them (since
     // there are multiple ways to make requests in Node).
-    // We patch once and once only, because otherwise
-    // it's challenging to test the server under different
-    // conditions.
     const getAppHost = () => options.appHostname
     http.request = outgoingRequestHook(http.request, getAppHost)
     http.get = outgoingRequestHook(http.get, getAppHost)