+++ packages/pwa-kit-react-sdk/CHANGELOG.md
@@ -1,6 +1,12 @@
-## To be released
+## v1.3.0-dev (Nov 18, 2021)
 
--   Upgrade `copy-webpack-plugin` to latest `^9.0.1` version. [3191](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/181)
+-   `createApp` takes a new option `enableLegacyRemoteProxying` which defaults to `true`. When set to `false`, local development proxying is disabled when running remotely. In future, local development proxying will *always* be disabled when running remotely. [#205](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/205)
+
+## v1.2.0 (Nov 18, 2021)
+
+-   Security package updates
+-   Upgrade `copy-webpack-plugin` to latest `^9.0.1` version. [#3191](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/181)
+
 ## v1.1.0 (Sep 27, 2021)
 
 -   Update the bundle push command to remove legacy bundle upload preview URL from console output. [#81](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/81)